ensure flight numbers csv exists from a variety of directories, etc

+++ ensure_flightnumbers_csv_exists.sh
@@ -1,4 +1,7 @@
 #!/bin/bash
 # encoding: utf-8
-mv FlightNumbers.csv FlightNumbers.csv.bak
-wget -q -O FlightNumbers.csv -nc http://www.virtualradarserver.co.uk/Files/FlightNumbers.csv
+
+if [ -f FlightNumbers.csv ]; then
+  mv FlightNumbers.csv FlightNumbers.csv.bak
+fi
+wget -q -O $(dirname $0)/FlightNumbers.csv -nc http://www.virtualradarserver.co.uk/Files/FlightNumbers.csv
